feat: add modal create idea body folder
