+++ src/components/Modals/ModalCreateIdea/ModalCreateIdeaBody/styles.ts
